refactor: 위도, 경도 double 타입 변경으로 인한 리팩토링

+++ src/main/java/com/example/fashionforecastbackend/region/domain/repository/RegionRepository.java
@@ -16,5 +16,5 @@ public interface RegionRepository extends JpaRepository<Region, Long>, RegionCus
 	Optional<Region> findByAddress(@Param("city") String city, @Param("district") String district,
 		@Param("neighborhood") String neighborhood);
 
-	Optional<Region> findByNxAndNy(int nx, int ny);
+	Optional<Region> findByNxAndNy(double nx, double ny);
 }
